refactor(solvers): split solvers.py into the layered package
numerics / etd2 / path / schedule + backends/{base,host,mkl,cuda,accelerate},
11 modules, largest 426 lines, none over the 600 limit.

Pure relocation — every module is assembled from line slices of the original
and an AST diff over all 85 defs and classes reports no executable change.
The fp64 solve is bitwise identical to 9afc10e on numpy and MKL, single-axis
and K=8, theta 0 and 60: 0 of 220968 cells differ. All six numerical goldens
pass untouched.

- The private surface narrows: 31 private helpers and 9 incidental import
  re-exports are no longer attributes of MCEq.solvers. Nothing in the tree
  imported them from there.
- tests/test_solvers_import.py pins that importing the package does not load
  cupy, dlopen libmkl, import MCEq.spacc or resolve a platform probe.
- .importlinter C5 respells the solvers -> config debt as the two edges the
  package now carries (backends.mkl for the MKL runtime, path for the step
  defaults); the MKL extraction removes the first.
- The C sources stay put. Moving them under solvers/_kernels changes the CMake
  install paths and the macOS build cannot be verified here — deferred to
  Phase 7, and the .importlinter comment now says so.
