+++ src/MCEq/solvers/numerics.py
@@ -1,0 +1,245 @@
+"""Scalar maths of the ETD2RK step: phi functions and step buffers.
+
+Everything the step formulas need that is neither a backend nor a policy:
+the two phi-function Taylor thresholds, the scratch-buffer layouts every
+backend allocates once per solve, the in-place diagonal-factor kernels
+(shared path and per-RHS path), the block phi factors of the sec(theta)
+exact slot, and the mode-space matmul.
+
+numpy only. A new scalar formula of the step loop belongs here; anything
+that touches a library handle, a device, or a configuration value does not.
+"""
+
+import numpy as np
+
+# phi1(z) = (e^z - 1) / z              (limit 1   as z -> 0)
+# phi2(z) = (e^z - 1 - z) / z**2       (limit 1/2 as z -> 0)
+# Below the analytic-formula cutoffs we patch with the order-2 Taylor
+# expansion via Horner. phi2 cancels at a wider radius than phi1 because
+# its numerator has a leading z² term.
+_PHI1_SMALL = 1e-6
+_PHI2_SMALL = 1e-3
+_INV_6 = 1.0 / 6.0
+_INV_24 = 1.0 / 24.0
+
+
+def _etd_step_buffers(dim):
+    """Allocate the per-step scratch arrays the ETD kernels need.
+
+    Centralized here so every backend shares an identical layout — this is
+    a hot loop, and any allocation inside it dominates the SpMVs once those
+    are running on Accelerate / MKL / a tuned BLAS.
+    """
+    return {
+        "D": np.empty(dim, dtype=np.float64),
+        "hD": np.empty(dim, dtype=np.float64),
+        "eD": np.empty(dim, dtype=np.float64),
+        "phi1": np.empty(dim, dtype=np.float64),
+        "phi2": np.empty(dim, dtype=np.float64),
+        "scratch": np.empty(dim, dtype=np.float64),
+        "abs_hD": np.empty(dim, dtype=np.float64),
+        "mask1": np.empty(dim, dtype=bool),
+        "mask2": np.empty(dim, dtype=bool),
+    }
+
+
+def _etd_step_buffers_multipath(dim, K):
+    """Scratch buffers for the per-RHS-path multi-RHS kernel.
+
+    Stage-3 lifts ``D`` / ``eD`` / ``φ₁`` / ``φ₂`` from ``(dim,)`` to
+    ``(dim, K)`` because both ``h`` and ``ri`` vary across columns (each
+    column carries its own atmosphere path). Memory cost at the full-sky
+    operating point dim=7986, K=3072, fp64: ≈ 600 MB for the three
+    (dim, K) diag buffers + (dim, K) scratch. See
+    wiki/methods/multi-rhs-etd2-design.md Stage 3 section.
+    """
+    return {
+        "D": np.empty((dim, K), dtype=np.float64),
+        "hD": np.empty((dim, K), dtype=np.float64),
+        "eD": np.empty((dim, K), dtype=np.float64),
+        "phi1": np.empty((dim, K), dtype=np.float64),
+        "phi2": np.empty((dim, K), dtype=np.float64),
+        "scratch": np.empty((dim, K), dtype=np.float64),
+        "abs_hD": np.empty((dim, K), dtype=np.float64),
+        "mask1": np.empty((dim, K), dtype=bool),
+        "mask2": np.empty((dim, K), dtype=bool),
+    }
+
+
+def _etd_compute_diag_factors_multipath(h_K, ri_K, d_int, d_dec, bufs):
+    """Per-RHS-path analogue of :func:`_etd_compute_diag_factors`.
+
+    Computes per-column ``D[i, k] = d_int[i] + ri_K[k] · d_dec[i]``,
+    ``hD = h_K[k] · D``, ``eD = exp(hD)``, and the two φ-functions of
+    ``hD`` elementwise over the ``(dim, K)`` plane. Branches around the
+    small-|hD| Taylor patch are computed via ``where=`` masks; numpy's
+    ufunc ``where=`` works on 2-D arrays so the per-cell branch is
+    cheap. cupy 14 rejects ``where=`` on most arithmetic ufuncs (verified
+    on the PriNCe port); when porting this to cupy, switch to the
+    ``copyto(where=)`` pattern used in PriNCe's etd2.py.
+
+    Frozen-column semantics: when ``h_K[k] == 0`` the column is "done"
+    (its own path has fewer steps than max). The math degenerates to
+    ``eD = 1, φ₁ = 1, φ₂ = 1/2`` for that column (Taylor branches at
+    hD = 0), and the downstream ETD2 update collapses to
+    ``state ← state``. No explicit masking needed.
+    """
+    D = bufs["D"]
+    hD = bufs["hD"]
+    eD = bufs["eD"]
+    phi1 = bufs["phi1"]
+    phi2 = bufs["phi2"]
+    scratch = bufs["scratch"]
+    abs_hD = bufs["abs_hD"]
+    mask1 = bufs["mask1"]
+    mask2 = bufs["mask2"]
+
+    # D[i, k] = d_int[i] + ri_K[k] · d_dec[i]  -- (dim, K), no extra alloc.
+    np.multiply(d_dec[:, None], ri_K[None, :], out=D)
+    np.add(D, d_int[:, None], out=D)
+    # hD = h_K[None, :] * D ; eD = exp(hD).
+    np.multiply(D, h_K[None, :], out=hD)
+    np.exp(hD, out=eD)
+
+    np.abs(hD, out=abs_hD)
+    np.greater(abs_hD, _PHI1_SMALL, out=mask1)
+    np.greater(abs_hD, _PHI2_SMALL, out=mask2)
+
+    # phi1: analytic (eD - 1) / hD where mask1, Taylor elsewhere.
+    np.subtract(eD, 1.0, out=phi1)
+    np.divide(phi1, hD, out=phi1, where=mask1)
+    np.multiply(hD, _INV_6, out=scratch)
+    np.add(scratch, 0.5, out=scratch)
+    np.multiply(scratch, hD, out=scratch)
+    np.add(scratch, 1.0, out=scratch)
+    np.invert(mask1, out=mask1)
+    np.copyto(phi1, scratch, where=mask1)
+
+    # phi2: analytic (eD - 1 - hD) / hD² where mask2, Taylor elsewhere.
+    np.subtract(eD, 1.0, out=phi2)
+    np.subtract(phi2, hD, out=phi2)
+    np.multiply(hD, hD, out=scratch)
+    np.divide(phi2, scratch, out=phi2, where=mask2)
+    np.multiply(hD, _INV_24, out=scratch)
+    np.add(scratch, _INV_6, out=scratch)
+    np.multiply(scratch, hD, out=scratch)
+    np.add(scratch, 0.5, out=scratch)
+    np.invert(mask2, out=mask2)
+    np.copyto(phi2, scratch, where=mask2)
+
+
+def _etd_compute_diag_factors(h, ri, d_int, d_dec, bufs):
+    """Fill ``bufs['eD']`` / ``bufs['phi1']`` / ``bufs['phi2']`` in place.
+
+    Computes the per-step diagonal of ``A + ri * B``, exponentiates it, and
+    evaluates the two phi-functions of ``h*D``. All work is done in
+    preallocated buffers — no temporaries — and the small-|hD| Taylor
+    branch is patched in only on the rows that need it (instead of being
+    evaluated eagerly across the whole array as ``np.where`` would).
+    """
+    D = bufs["D"]
+    hD = bufs["hD"]
+    eD = bufs["eD"]
+    phi1 = bufs["phi1"]
+    phi2 = bufs["phi2"]
+    scratch = bufs["scratch"]
+    abs_hD = bufs["abs_hD"]
+    mask1 = bufs["mask1"]
+    mask2 = bufs["mask2"]
+
+    # D = d_int + ri * d_dec
+    np.multiply(d_dec, ri, out=D)
+    np.add(D, d_int, out=D)
+    # hD = h * D ; eD = exp(hD)
+    np.multiply(D, h, out=hD)
+    np.exp(hD, out=eD)
+
+    # Branch masks: True ⇒ analytic form is safe.
+    np.abs(hD, out=abs_hD)
+    np.greater(abs_hD, _PHI1_SMALL, out=mask1)
+    np.greater(abs_hD, _PHI2_SMALL, out=mask2)
+
+    # phi1: analytic (eD - 1) / hD where mask1, Taylor 1 + hD/2 + hD²/6 elsewhere.
+    np.subtract(eD, 1.0, out=phi1)
+    np.divide(phi1, hD, out=phi1, where=mask1)
+    # Horner Taylor for phi1: ((hD/6) + 1/2)*hD + 1
+    np.multiply(hD, _INV_6, out=scratch)
+    np.add(scratch, 0.5, out=scratch)
+    np.multiply(scratch, hD, out=scratch)
+    np.add(scratch, 1.0, out=scratch)
+    np.invert(mask1, out=mask1)  # mask1 now: small |hD| rows
+    np.copyto(phi1, scratch, where=mask1)
+
+    # phi2: analytic (eD - 1 - hD) / hD² where mask2, Taylor 1/2 + hD/6 + hD²/24 elsewhere.
+    np.subtract(eD, 1.0, out=phi2)
+    np.subtract(phi2, hD, out=phi2)
+    np.multiply(hD, hD, out=scratch)  # hD² in scratch
+    np.divide(phi2, scratch, out=phi2, where=mask2)
+    # Horner Taylor for phi2: ((hD/24) + 1/6)*hD + 1/2
+    np.multiply(hD, _INV_24, out=scratch)
+    np.add(scratch, _INV_6, out=scratch)
+    np.multiply(scratch, hD, out=scratch)
+    np.add(scratch, 0.5, out=scratch)
+    np.invert(mask2, out=mask2)  # mask2 now: small |hD| rows
+    np.copyto(phi2, scratch, where=mask2)
+
+
+def _secant_phi_factors(ZB, out=None, work=None):
+    """Elementwise exp/phi1/phi2 with Taylor patches for a block argument.
+
+    ``out=(exp, phi1, phi2)`` and ``work=(scratch, large)`` let the hot
+    loop reuse its block-sized arrays instead of allocating six
+    temporaries per step. The no-argument form keeps the small standalone
+    helper convenient for tests and callers outside the driver.
+    """
+    if out is not None:
+        eDB, phi1, phi2 = out
+        scratch, large = work
+        np.expm1(ZB, out=scratch)
+
+        # Taylor branches, evaluated for the whole block and overwritten
+        # by the quotient on entries outside the small-argument patch.
+        np.multiply(ZB, _INV_6, out=phi1)
+        np.add(phi1, 0.5, out=phi1)
+        np.multiply(ZB, phi1, out=phi1)
+        np.add(phi1, 1.0, out=phi1)
+        np.greater(np.abs(ZB, out=eDB), _PHI1_SMALL, out=large)
+        np.divide(scratch, ZB, out=phi1, where=large)
+
+        np.multiply(ZB, _INV_24, out=phi2)
+        np.add(phi2, _INV_6, out=phi2)
+        np.multiply(ZB, phi2, out=phi2)
+        np.add(phi2, 0.5, out=phi2)
+        np.greater(eDB, _PHI2_SMALL, out=large)
+        np.subtract(scratch, ZB, out=eDB)
+        np.multiply(ZB, ZB, out=scratch)
+        np.divide(eDB, scratch, out=phi2, where=large)
+        np.exp(ZB, out=eDB)
+        return out
+
+    safe = np.where(ZB == 0.0, 1.0, ZB)
+    e1 = np.expm1(ZB)
+    phi1 = np.where(np.abs(ZB) > _PHI1_SMALL, e1 / safe, 1.0 + ZB * (0.5 + ZB * _INV_6))
+    phi2 = np.where(
+        np.abs(ZB) > _PHI2_SMALL,
+        (e1 - ZB) / (safe * safe),
+        0.5 + ZB * (_INV_6 + ZB * _INV_24),
+    )
+    return np.exp(ZB), phi1, phi2
+
+
+def _secant_left_matmul(matrix, plane, out=None):
+    """Apply a mode-space matrix to the leading axis of a >= 2-D plane.
+
+    The driver carries logical ``(mode, column, lane)`` planes; flattening
+    the trailing axes turns each mode transform into one dense GEMM. The
+    plane may be a strided view of the state (the low-E block) — BLAS
+    takes it through its leading dimension without a copy.
+    """
+    trailing = plane.shape[1:]
+    result_shape = (matrix.shape[0],) + trailing
+    plane_2d = plane.reshape(plane.shape[0], -1)
+    if out is None:
+        return (matrix @ plane_2d).reshape(result_shape)
+    np.matmul(matrix, plane_2d, out=out.reshape(matrix.shape[0], -1))
+    return out
