+++ src/MCEq/solvers/path.py
@@ -1,0 +1,218 @@
+"""The integration path the ETD2 kernels step along.
+
+:func:`etd2_nonuniform_path` turns a density model into the
+``(nsteps, dX, rho_inv, grid_idcs)`` tuple of the kernel-dispatch contract:
+step sizes from the local variation of ``1/rho``, and ``rho_inv`` as the
+per-step integral mean rather than a point sample. The module also carries
+the documented EM-row caveat the kernels tolerate.
+
+Step-size policy and the depth grid belong here. The step formulas are in
+:mod:`MCEq.solvers.numerics` and the loop that consumes the path is
+:func:`MCEq.solvers.etd2.etd2_driver`.
+"""
+
+import numpy as np
+
+#: Documented contract for the EM-row blowup at extreme zenith. Referenced
+#: from each ETD2 kernel; see ``docs/mceq_v1.x_v2_diff.md`` "EM cascade
+#: caveat" for the full derivation. Until a validated EM database ships,
+#: ``config.adv_set["disabled_particles"]`` defaults to ``[11, -11]`` so
+#: this branch is never entered for production runs.
+_EM_BLOWUP_CAVEAT = """\
+At extreme zenith the e± semi-Lagrangian L/R rows produce ``inf`` in
+``F_phi`` / ``F_a`` (no diagonal damping). The blowup is contained to
+those rows: e±/γ do not feed back into hadrons via ``int_m`` / ``dec_m``,
+so muons and neutrinos are unaffected. Each ETD2 kernel wraps its loop
+with ``np.errstate(over='ignore', invalid='ignore')`` to suppress the
+resulting overflow / NaN warnings. To exclude the EM block entirely, set
+``config.adv_set['disabled_particles'] = [11, -11]`` (the default).
+"""
+
+
+def etd2_nonuniform_path(
+    density_model,
+    *,
+    X_start=None,
+    eps=None,
+    dX_max=None,
+    dX_min=None,
+    fd_span=None,
+    int_grid=None,
+    step=None,
+):
+    """Build a non-uniform integration path for ETD2 kernels.
+
+    Step sizes follow ``h_k = min(dX_max, eps / |d ln rho_inv / dX|(X_k))``
+    with a ``max(.., dX_min)`` floor. ``rho_inv`` for the kernel is the
+    integral mean of ``density_model.r_X2rho`` over each step (via
+    ``scipy.integrate.quad``), not a point sample — this is essential at
+    the very first step which crosses the spline-saturation cap.
+
+    See ``docs/mceq_v1.x_v2_diff.md`` ("Step-size control and the resonance
+    approximation") for the design.
+
+    Args:
+      density_model: object with ``r_X2rho(X)`` and ``max_X``.
+      X_start (float | None): starting depth in g/cm^2; ``None`` →
+        ``step.X_start``.
+      eps (float | None): within-step ``rho_inv`` variation tolerance;
+        ``None`` → ``step.etd2_path["eps"]``.
+      dX_max (float | None): cap on step size (off-diagonal stability
+        cliff); ``None`` → ``step.etd2_path["dX_max"]``.
+      dX_min (float | None): floor on step size; ``None`` →
+        ``step.etd2_path["dX_min"]``.
+      fd_span (float | None): forward-FD probe span; ``None`` →
+        ``step.etd2_path["fd_span"]``.
+      int_grid (np.ndarray | None): X values at which to record snapshots.
+        Steps are truncated to land exactly on each ``int_grid`` entry.
+      step: the ``solver`` setting group (``X_start``, ``etd2_path``);
+        ``None`` → ``MCEq.config.solver``. Read here, when the path is
+        planned — a live instance's next ``solve()`` does not see a write
+        made after the path was built, because
+        ``MCEqRun._calculate_integration_path`` caches on the unresolved
+        keyword arguments.
+
+    Returns:
+      (nsteps, dX, rho_inv, grid_idcs): tuple compatible with the
+      kernel-dispatch contract used by ``MCEqRun.integration_path``.
+    """
+    if step is None:
+        from MCEq import config
+
+        step = config.solver
+    if X_start is None:
+        X_start = step.X_start
+    p = step.etd2_path
+    if eps is None:
+        eps = p["eps"]
+    if dX_max is None:
+        dX_max = p["dX_max"]
+    if dX_min is None:
+        dX_min = p["dX_min"]
+    if fd_span is None:
+        fd_span = p["fd_span"]
+
+    ri = density_model.r_X2rho
+    max_X = density_model.max_X
+    n_int = int(np.size(int_grid)) if int_grid is not None else 0
+
+    if n_int and float(np.min(int_grid)) < float(X_start):
+        raise ValueError(
+            "Steps in int_grid must be larger than or equal to X_start "
+            f"(got min(int_grid)={float(np.min(int_grid)):.6g}, "
+            f"X_start={float(X_start):.6g})."
+        )
+
+    Xs, dXs, grid_idcs = [], [], []
+    grid_step = 0
+    X = float(X_start)
+    step = 0
+    while X < max_X:
+        rate = abs(np.log(float(ri(X + fd_span))) - np.log(float(ri(X)))) / fd_span
+        h = min(dX_max, eps / rate) if rate > 0 else dX_max
+        h = max(h, dX_min)
+        h = min(h, max_X - X)
+        # Truncate the step to land exactly on the next user-requested
+        # snapshot point. The truncation can drive h below dX_min — that's
+        # the user's intent, not a stability issue (smaller h is always
+        # more stable for ETD2). It can chain across many short steps if
+        # the user's grid is finer than the natural schedule.
+        if n_int and grid_step < n_int and X + h >= int_grid[grid_step]:
+            h = float(int_grid[grid_step]) - X
+            grid_idcs.append(step)
+            grid_step += 1
+        Xs.append(X)
+        dXs.append(h)
+        X += h
+        step += 1
+
+    Xs = np.asarray(Xs, dtype=np.float64)
+    dXs = np.asarray(dXs, dtype=np.float64)
+    rho_inv = np.empty(len(dXs), dtype=np.float64)
+
+    # Compute per-step integral means of ``r_X2rho`` via differences on
+    # a cumulative-trapezoid antiderivative built once on a hybrid
+    # log+linear sample.
+    #
+    # Why a hybrid sample (and not a uniform one): ``r_X2rho`` is
+    # ``1/rho``, and atmosphere density splines deliberately saturate at
+    # the top of atmosphere so that the path-builder's FD probe stays
+    # well-defined. The saturation produces a step-function-like spike
+    # near ``X = 0``: ``ri(0) ~ 1e9 cm^3/g`` falls to ``~1e7`` by
+    # ``X = 0.01``. A uniform-grid quadrature smears that spike across
+    # samples and over-estimates the mean by ~2-20x for the first few
+    # steps. Sampling logarithmically near ``X = 0`` (concentrated where
+    # ``ri`` varies fast) and linearly in the bulk recovers
+    # ``quad(epsrel=1e-6)``-class accuracy at ``O(1)`` per step.
+    #
+    # Why a spline of the *cumulative* and not of ``r_X2rho`` directly:
+    # the cumulative is smooth and strictly monotone, so a cubic spline
+    # through it is well-behaved even though ``r_X2rho`` spans 5+ decades.
+    # A direct fit overshoots and yields non-physical (negative) means
+    # near the top of atmosphere.
+    #
+    # Falls back to ``quad`` only if ``ri`` rejects array input.
+    use_cum = False
+    cum_spline = None
+    if len(dXs) > 0:
+        try:
+            from scipy.integrate import cumulative_trapezoid
+            from scipy.interpolate import UnivariateSpline
+
+            X_min = float(Xs[0])
+            X_max = float(Xs[-1] + dXs[-1])
+            # Cap at the model's max depth (numerical drift on the last step).
+            X_max = min(X_max, float(max_X))
+            if X_max > X_min:
+                # Concentrated log-sample plus a dense linear sample for
+                # the bulk. The log part must span the WHOLE domain, not
+                # stop at X = 1: on near-horizontal trajectories
+                # (max_X ~ 3e4) the interval X = 1..50 g/cm^2 still sits
+                # at 80..35 km altitude where 1/rho varies over decades,
+                # and a linear bulk sample alone (max_X/4000 ~ 7 g/cm^2
+                # per point) under-resolves it. The mis-integrated
+                # per-step means (x0.4..x1.6) imprinted a coherent
+                # 10-25% bump-dip on ALL species at X ~ 5-40 g/cm^2 for
+                # theta >~ 85 deg (found via nu3d q-table wiggles at
+                # ~40 km altitude, 2026-07-10). Log sampling costs
+                # nothing in the bulk, where 1/rho is flat per log-X;
+                # the linear sample is kept so vertical-path accuracy
+                # is unchanged.
+                X_log_lo = max(1e-7, X_min if X_min > 0 else 1e-7)
+                if X_max > X_log_lo:
+                    X_top = np.geomspace(X_log_lo, X_max, 6001)
+                else:
+                    X_top = np.empty(0)
+                X_bulk = np.linspace(max(X_log_lo, X_min), X_max, 4001)
+                sample_X = np.unique(np.r_[X_min, X_top, X_bulk])
+                sample_X.sort()
+                sample_ri = np.asarray(ri(sample_X), dtype=np.float64)
+                if (
+                    sample_ri.shape == sample_X.shape
+                    and np.all(np.isfinite(sample_ri))
+                    and np.all(sample_ri > 0.0)
+                ):
+                    cum = cumulative_trapezoid(sample_ri, sample_X, initial=0.0)
+                    cum_spline = UnivariateSpline(sample_X, cum, k=3, s=0.0)
+                    use_cum = True
+        except Exception:
+            use_cum = False
+
+    if not use_cum:
+        from scipy.integrate import quad
+
+    for i in range(len(dXs)):
+        # A zero-length step occurs when ``int_grid`` requests a snapshot
+        # at X_start: the truncation drives the first dX to 0 and the kernel
+        # records the initial state. Point-sample ri there to avoid /0.
+        if dXs[i] == 0.0:
+            rho_inv[i] = float(ri(Xs[i]))
+        elif use_cum:
+            a = Xs[i]
+            b = a + dXs[i]
+            rho_inv[i] = float(cum_spline(b) - cum_spline(a)) / dXs[i]
+        else:
+            rho_inv[i] = (
+                quad(ri, Xs[i], Xs[i] + dXs[i], limit=50, epsrel=1e-6)[0] / dXs[i]
+            )
+    return len(dXs), dXs, rho_inv, grid_idcs
