+++ src/MCEq/solvers/schedule.py
@@ -1,0 +1,176 @@
+"""LPT carousel schedule for the multi-RHS route.
+
+``K_total`` pixels stream through a fixed-width ``K`` pipeline; when a slot
+finishes its current pixel, the next pixel's phi0 is loaded into that
+slot's column on the same step. :func:`schedule_lpt` assigns pixels to
+slots and :func:`compile_carousel_schedule` turns that assignment into the
+``(T, K)`` path tensors and the sparse reset / record events of a
+:class:`CarouselSchedule`. Pure numpy and backend-agnostic;
+:func:`MCEq.solvers.etd2.etd2_driver` consumes the schedule as stage 7
+(harvest before reset) of its step loop.
+
+Packing policy belongs here -- how pixels share a pipeline. What the loop
+does with a schedule belongs in :mod:`MCEq.solvers.etd2`.
+
+Design: ../mceq-em-integration/wiki/methods/multi-rhs-lpt-carousel.md
+"""
+
+from collections import namedtuple
+
+import numpy as np
+
+CarouselSchedule = namedtuple(
+    "CarouselSchedule",
+    [
+        "T",  # int — makespan (outer loop iters)
+        "K",  # int — pipeline width (slots)
+        "K_total",  # int — total pixels packed
+        "slot_assignments",  # list[list[int]] — per-slot pixel ids in run order
+        "reset_t_starts",  # (T+1,) int32 — CSR ptrs into reset_j / reset_pixel
+        "reset_j",  # (R,) int32 — slot id of each reset event
+        "reset_pixel",  # (R,) int32 — pixel id whose phi0 to load
+        "record_t_starts",  # (T+1,) int32 — CSR ptrs into record_j / record_pixel
+        "record_j",  # (K_total,) int32 — slot id of each harvest event
+        "record_pixel",  # (K_total,) int32 — pixel id to record into
+    ],
+)
+
+
+def schedule_lpt(nsteps_per_pixel, K):
+    """LPT (longest-processing-time-first) multiway-makespan assignment.
+
+    Sorts pixels by ``nsteps`` descending and greedily appends each to the
+    slot with the currently smallest running length sum. LPT is guaranteed
+    to be within 4/3 of optimal; in our regime (no single pixel
+    dominates) it typically achieves ``T ≈ ⌈Σ/K⌉``.
+
+    Args:
+        nsteps_per_pixel: array-like of int, length K_total.
+        K: int — desired pipeline width. Clamped to ``min(K, K_total)``.
+
+    Returns:
+        slot_assignments: list of K lists; slot j → ordered pixel ids.
+        T: int — makespan = max over slots of total assigned nsteps.
+
+    Notes:
+        Pixel order within a slot does not affect the makespan; we keep
+        the natural LPT order (longest first) for determinism.
+    """
+    import heapq
+
+    ns = np.asarray(nsteps_per_pixel, dtype=np.int64)
+    K_total = int(ns.size)
+    K_eff = int(min(K, K_total))
+    if K_eff < 1:
+        raise ValueError(f"schedule_lpt: K must be >= 1 (got {K})")
+
+    order = np.argsort(ns, kind="stable")[::-1]  # longest first
+
+    # Min-heap keyed on (current slot length, slot id). The list of pixel
+    # ids per slot lives outside the heap to keep heap entries small.
+    heap = [(0, j) for j in range(K_eff)]
+    heapq.heapify(heap)
+    slot_assignments = [[] for _ in range(K_eff)]
+    for pid in order:
+        pid_i = int(pid)
+        L_j, j = heapq.heappop(heap)
+        slot_assignments[j].append(pid_i)
+        heapq.heappush(heap, (L_j + int(ns[pid_i]), j))
+
+    T = max(int(ns[s].sum() if s else 0) for s in slot_assignments) if True else 0
+    # Recompute T cleanly from the heap residuals:
+    T = max(L_j for L_j, _ in heap)
+    return slot_assignments, T
+
+
+def compile_carousel_schedule(paths, slot_assignments, T, dim, phi0_per_pixel):
+    """Build the (T, K) path tensors and sparse reset/record events.
+
+    Concatenates each slot's pixel paths end-to-end into columns of
+    ``dX_2d`` / ``rho_inv_2d``. Records the per-pixel harvest step (last
+    step of that pixel's slice within its slot) and the per-pixel reset
+    step (right after the prior pixel's harvest, except for the first
+    pixel in a slot which is loaded directly into ``phi_initial``).
+
+    Args:
+        paths: list of ``(nsteps, dX_k, rho_inv_k, _grid_idcs)`` tuples,
+            indexed by pixel id.
+        slot_assignments: from :func:`schedule_lpt`.
+        T: makespan from :func:`schedule_lpt`.
+        dim: state dimension.
+        phi0_per_pixel: ``(dim, K_total)`` array — per-pixel initial phi.
+
+    Returns:
+        dX_carousel: ``(T, K)`` f64 — slot-concatenated step sizes,
+            zero-padded after each slot's total length.
+        rho_inv_carousel: ``(T, K)`` f64 — slot-concatenated densities.
+        phi_initial: ``(dim, K)`` f64 — first pixel's phi0 per slot.
+        schedule: :class:`CarouselSchedule`.
+    """
+    K = len(slot_assignments)
+    K_total = sum(len(s) for s in slot_assignments)
+
+    dX_2d = np.zeros((T, K), dtype=np.float64)
+    rho_inv_2d = np.zeros((T, K), dtype=np.float64)
+    phi_initial = np.zeros((dim, K), dtype=np.float64)
+
+    reset_per_t = [[] for _ in range(T)]
+    record_per_t = [[] for _ in range(T)]
+
+    for j, pixels in enumerate(slot_assignments):
+        if not pixels:
+            continue
+        phi_initial[:, j] = phi0_per_pixel[:, pixels[0]]
+        t_cursor = 0
+        for i, pid in enumerate(pixels):
+            ns_p, dX_p, ri_p, _ = paths[pid]
+            if int(ns_p) != len(dX_p) or int(ns_p) != len(ri_p):
+                raise ValueError(
+                    f"compile_carousel_schedule: pixel {pid} path "
+                    f"length mismatch (nsteps={ns_p}, len(dX)={len(dX_p)}, "
+                    f"len(rho_inv)={len(ri_p)})"
+                )
+            dX_2d[t_cursor : t_cursor + ns_p, j] = dX_p
+            rho_inv_2d[t_cursor : t_cursor + ns_p, j] = ri_p
+            t_finish = t_cursor + ns_p - 1
+            record_per_t[t_finish].append((j, pid))
+            t_cursor += ns_p
+            if i + 1 < len(pixels):
+                reset_per_t[t_finish].append((j, pixels[i + 1]))
+
+    reset_t_starts = np.zeros(T + 1, dtype=np.int32)
+    record_t_starts = np.zeros(T + 1, dtype=np.int32)
+    for t in range(T):
+        reset_t_starts[t + 1] = reset_t_starts[t] + len(reset_per_t[t])
+        record_t_starts[t + 1] = record_t_starts[t] + len(record_per_t[t])
+    R = int(reset_t_starts[T])
+    Rec = int(record_t_starts[T])
+    reset_j = np.empty(R, dtype=np.int32)
+    reset_pixel = np.empty(R, dtype=np.int32)
+    record_j = np.empty(Rec, dtype=np.int32)
+    record_pixel = np.empty(Rec, dtype=np.int32)
+    r_c = 0
+    rec_c = 0
+    for t in range(T):
+        for j, pid in reset_per_t[t]:
+            reset_j[r_c] = j
+            reset_pixel[r_c] = pid
+            r_c += 1
+        for j, pid in record_per_t[t]:
+            record_j[rec_c] = j
+            record_pixel[rec_c] = pid
+            rec_c += 1
+
+    schedule = CarouselSchedule(
+        T=T,
+        K=K,
+        K_total=K_total,
+        slot_assignments=slot_assignments,
+        reset_t_starts=reset_t_starts,
+        reset_j=reset_j,
+        reset_pixel=reset_pixel,
+        record_t_starts=record_t_starts,
+        record_j=record_j,
+        record_pixel=record_pixel,
+    )
+    return dX_2d, rho_inv_2d, phi_initial, schedule
